Backend/requirements.txt
python-dotenv
gunicorn==20.1.0
Pillow
psycopg2-binary
djangorestframework
django
django-cors-headers
